# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: GCLI2API Gateway API
  version: 0.1.0
  description: |
    OpenAPI 规范覆盖 GCLI2API-Go 暴露的对外接口，包括：
      * OpenAI 兼容端点 `/v1/*`
      * Gemini 原生端点 `/v1` 与 `/v1beta`
      * 管理端 `/routes/api/management/*`
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://{host}{basePath}
    description: 生产或预发环境
    variables:
      host:
        default: gateway.internal
      basePath:
        default: ""
  - url: http://{devHost}
    description: 本地或测试环境
    variables:
      devHost:
        default: dev-gateway.internal:8080
security:
  - ApiKeyAuth: []
  - GoogleApiKey: []
  - XApiKey: []
paths:
  /v1/models:
    get:
      tags: [OpenAI Compatibility]
      summary: 列出可用模型
      operationId: listOpenAIModels
      description: 返回 OpenAI 兼容模型清单。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIModelsResponse'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/models/{id}:
    get:
      tags: [OpenAI Compatibility]
      summary: 获取模型详情
      operationId: getOpenAIModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 模型 ID。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIModelResponse'
        '404':
          description: 未找到
        '400':
          $ref: '#/components/responses/Error400'
  /v1/chat/completions:
    post:
      tags: [OpenAI Compatibility]
      summary: 创建聊天补全
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionsRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionsResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/completions:
    post:
      tags: [OpenAI Compatibility]
      summary: 创建文本补全
      operationId: createLegacyCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyCompletionsRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegacyCompletionsResponse'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/responses:
    post:
      tags: [OpenAI Compatibility]
      summary: 创建多模态响应
      operationId: createResponses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsesResponse'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/images/generations:
    post:
      tags: [OpenAI Compatibility]
      summary: 生成图像
      operationId: generateImages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImagesGenerationRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagesGenerationResponse'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/models/{model}:generateContent:
    post:
      tags: [Gemini Native]
      summary: 生成 Gemini 响应
      operationId: generateGeminiContent
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeminiGenerateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeminiGenerateResponse'
        '400':
          $ref: '#/components/responses/Error400'
  /v1/models/{model}:streamGenerateContent:
    post:
      tags: [Gemini Native]
      summary: Gemini 流式生成
      operationId: streamGeminiContent
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeminiGenerateRequest'
      responses:
        '200':
          description: 事件流
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Error400'
  /v1/models/{model}:countTokens:
    post:
      tags: [Gemini Native]
      summary: 计算 Token
      operationId: countGeminiTokens
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeminiCountTokensRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeminiCountTokensResponse'
        '400':
          $ref: '#/components/responses/Error400'
...
components:
  schemas:
    OpenAIModelsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OpenAIModel'
        object:
          type: string
    OpenAIModelResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OpenAIModel'
    OpenAIModel:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
          format: int64
        owned_by:
          type: string
        capabilities:
          type: object
          additionalProperties: true
    ChatCompletionsRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        temperature:
          type: number
        top_p:
          type: number
        max_tokens:
          type: integer
        stream:
          type: boolean
        response_format:
          type: object
          additionalProperties: true
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        name:
          type: string
    ChatCompletionsResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
          format: int64
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              finish_reason:
                type: string
              message:
                $ref: '#/components/schemas/ChatMessage'
        usage:
          $ref: '#/components/schemas/TokenUsage'
    TokenUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    LegacyCompletionsRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        max_tokens:
          type: integer
        temperature:
          type: number
        stream:
          type: boolean
    LegacyCompletionsResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
        choices:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              index:
                type: integer
              finish_reason:
                type: string
        usage:
          $ref: '#/components/schemas/TokenUsage'
    ResponsesRequest:
      type: object
      properties:
        model:
          type: string
        input:
          type: array
          items:
            type: object
            additionalProperties: true
    ResponsesResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        output:
          type: array
          items:
            type: object
            additionalProperties: true
    ImagesGenerationRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        size:
          type: string
    ImagesGenerationResponse:
      type: object
      properties:
        created:
          type: integer
        data:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
    GeminiGenerateRequest:
      type: object
      properties:
        contents:
          type: array
          items:
            $ref: '#/components/schemas/GeminiContent'
        safetySettings:
          type: array
          items:
            type: object
            additionalProperties: true
    GeminiContent:
      type: object
      properties:
        role:
          type: string
        parts:
          type: array
          items:
            type: object
            additionalProperties: true
    GeminiGenerateResponse:
      type: object
      properties:
        candidates:
          type: array
          items:
            type: object
            properties:
              content:
                $ref: '#/components/schemas/GeminiContent'
              finishReason:
                type: string
        usageMetadata:
          type: object
          additionalProperties: true
    GeminiCountTokensRequest:
      type: object
      properties:
        contents:
          type: array
          items:
            $ref: '#/components/schemas/GeminiContent'
    GeminiCountTokensResponse:
      type: object
      properties:
        totalTokens:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
  responses:
    Error400:
      description: 请求无效或参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 接受 `Bearer <token>` 或直接提供预共享 Token。
    GoogleApiKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: Gemini 兼容头部。
    XApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Anthropic 风格 API Key 头部。